+++ classes/activities/class-activity.php
@@ -5,7 +5,7 @@
  * @package Progress_Planner
  */
 
-namespace Progress_Planner;
+namespace Progress_Planner\Activities;
 
 /**
  * Activity class.
