+++ classes/class-query.php
@@ -192,7 +192,7 @@ class Query {
 	/**
 	 * Insert multiple activities into the database.
 	 *
-	 * @param \Progress_Planner\Activity[] $activities The activities to insert.
+	 * @param \Progress_Planner\Activities\Activity[] $activities The activities to insert.
 	 *
 	 * @return int[]|false The IDs of the inserted activities, or false on failure.
 	 */
@@ -215,7 +215,7 @@ class Query {
 	/**
 	 * Insert an activity into the database.
 	 *
-	 * @param \Progress_Planner\Activity $activity The activity to insert.
+	 * @param \Progress_Planner\Activities\Activity $activity The activity to insert.
 	 *
 	 * @return int|false The ID of the inserted activity, or false on failure.
 	 */
@@ -255,7 +255,7 @@ class Query {
 	 *
 	 * @param array $results The results.
 	 *
-	 * @return \Progress_Planner\Activity[] The activities.
+	 * @return \Progress_Planner\Activities\Activity[] The activities.
 	 */
 	private function get_activities_from_results( $results ) {
 		$activities = [];
@@ -277,8 +277,8 @@ class Query {
 	/**
 	 * Update an activity in the database.
 	 *
-	 * @param int                        $id       The ID of the activity to update.
-	 * @param \Progress_Planner\Activity $activity The activity to update.
+	 * @param int                                   $id       The ID of the activity to update.
+	 * @param \Progress_Planner\Activities\Activity $activity The activity to update.
 	 *
 	 * @return void
 	 */
@@ -312,7 +312,7 @@ class Query {
 	/**
 	 * Delete activities from the database.
 	 *
-	 * @param \Progress_Planner\Activity[] $activities The activity to delete.
+	 * @param \Progress_Planner\Activities\Activity[] $activities The activity to delete.
 	 *
 	 * @return void
 	 */
@@ -325,7 +325,7 @@ class Query {
 	/**
 	 * Delete an activity from the database.
 	 *
-	 * @param \Progress_Planner\Activity $activity The activity to delete.
+	 * @param \Progress_Planner\Activities\Activity $activity The activity to delete.
 	 *
 	 * @return void
 	 */
@@ -378,7 +378,7 @@ class Query {
 	 *
 	 * @param int $limit The number of activities to get.
 	 *
-	 * @return \Progress_Planner\Activity[]|null The activities. Returns null if there are no activities.
+	 * @return \Progress_Planner\Activities\Activity[]|null The activities. Returns null if there are no activities.
 	 */
 	public function get_latest_activities( $limit = 5 ) {
 		global $wpdb;
@@ -402,7 +402,7 @@ class Query {
 	/**
 	 * Get oldest activity.
 	 *
-	 * @return \Progress_Planner\Activity|null Returns null if there are no activities.
+	 * @return \Progress_Planner\Activities\Activity|null Returns null if there are no activities.
 	 */
 	public function get_oldest_activity() {
 		global $wpdb;
@@ -442,7 +442,7 @@ class Query {
 		if ( class_exists( '\Progress_Planner\Activities\\' . ucfirst( $category ) ) ) {
 			return '\Progress_Planner\Activities\\' . ucfirst( $category );
 		}
-		return '\Progress_Planner\Activity';
+		return '\Progress_Planner\Activities\Activity';
 	}
 
 	/**
