+++ tests/phpunit/test-class-activity.php
@@ -7,7 +7,7 @@
 
 namespace Progress_Planner\Tests;
 
-use Progress_Planner\Activity;
+use Progress_Planner\Activities\Activity;
 
 /**
  * Date test case.
