+++ tests/phpunit/test-class-badges.php
@@ -7,9 +7,6 @@
 
 namespace Progress_Planner\Tests;
 
-use Progress_Planner\Activity;
-use Progress_Planner\Badges\Monthly;
-
 /**
  * Badges test case.
  */
