+++ tests/phpunit/test-class-monthly-badge.php
@@ -7,7 +7,7 @@
 
 namespace Progress_Planner\Tests;
 
-use Progress_Planner\Activity;
+use Progress_Planner\Activities\Activity;
 use Progress_Planner\Badges\Monthly;
 
 /**
